More flexibility in freezing parser world age
This separates the decision about whether to freeze the parser world age
from Core.Compiler.parse. The builtin _apply_in_world allows this to be
done with more flexibility from outside (for example, one might want to
do some argument conversion in the current world before calling through
to a fixed world for the parser internals.